ImClassApp/requirements.txt
absl-py==0.12.0
alabaster
appdirs==1.4.4
argh==0.26.2
astroid
astunparse==1.6.3
async-generator
atomicwrites==1.4.0
attrs
autopep8
Babel
backcall
black==19.10b0
bleach
Brotli
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.6.20
cffi
chardet
click==7.1.2
cloudpickle
colorama
cryptography
cycler==0.10.0
cytoolz==0.11.0
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-renderer
dash-table
dask
decorator
defusedxml
diff-match-patch
docutils==0.16
entrypoints==0.3
flake8
Flask==1.1.2
Flask-Compress
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna
imageio
imagesize
importlib-metadata
intervaltree
ipykernel
ipython
ipython-genutils
isort
itsdangerous==1.1.0
jedi
jeepney
Jinja2==2.11.2
joblib
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
Keras-Preprocessing==1.1.2
keyring
kiwisolver
lazy-object-proxy
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient
nbconvert
nbformat
nest-asyncio
networkx
numpy
numpydoc
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging
pandocfilters
parso==0.7.0
pathspec==0.7.0
pexpect
pickleshare
Pillow
plotly==4.13.0
pluggy
prompt-toolkit
protobuf==3.15.6
psutil
ptyprocess
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle
pycparser
pydocstyle
pyflakes
Pygments
pylint
pyls-black
pyls-spyder
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil==2.8.1
python-jsonrpc-server
python-language-server
pytz==2020.1
PyWavelets
pyxdg
PyYAML==5.3.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtAwesome
qtconsole
QtPy==1.9.0
regex
requests
requests-oauthlib==1.3.0
retrying==1.3.3
rope
rsa==4.7.2
scikit-image==0.17.2
scipy
SecretStorage
six
snowballstemmer
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu
tensorflow-estimator==2.4.0
termcolor==1.1.0
testpath
textdistance
threadpoolctl
three-merge
toml
toolz
tornado
traitlets
typed-ast
typing-extensions
ujson
urllib3
watchdog
wcwidth
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer
yapf
zipp

